feat: Add ScoreResponse
- Add GetScoresDto in ScoreResponse
- Remove GetScoresDto in PostResponse
